removing the parse api
I was already debating wether this unsafe method should exist.

rfc7519 says it should not:

When validating a JWT, the following steps are performed.  The order
of the steps is not significant in cases where there are no
dependencies between the inputs and outputs of the steps.  If any of
the listed steps fail, then the JWT MUST be rejected -- that is,
treated by the application as an invalid input.

I feel like it's best to remove it
