+++ index.js
@@ -197,9 +197,6 @@ Jwt.prototype.isExpired = function() {
 
 
 function Parser(options){
-  if(!(this instanceof Parser)){
-    return new Parser(options);
-  }
   return this;
 }
 
@@ -325,7 +322,6 @@ var jwtLib = {
   Jwt: Jwt,
   JwtBody: JwtBody,
   JwtHeader: JwtHeader,
-  Parser: Parser,
   Verifier: Verifier,
   base64urlEncode: base64urlEncode,
   base64urlUnescape:base64urlUnescape,
@@ -366,9 +362,6 @@ var jwtLib = {
     }
 
     return jwt;
-  },
-  parse: function parse(jwtString,cb){
-    return new Parser().parse(jwtString,cb);
   }
 };
 module.exports = jwtLib;
