+++ test/verifier.js
@@ -46,6 +46,19 @@ describe('.verify()',function(){
     });
   });
 
+  it('should return PARSE_ERROR if the header is not JSON',function(){
+    assert.throws(function(){
+      nJwt.verify("noavalidheader.notavalidbody");
+    },properties.errors.PARSE_ERROR);
+  });
+
+  it('should return PARSE_ERROR if the body is not JSON',function(){
+    var header = nJwt.JwtHeader({type:'JWT',alg:'HS256'}).compact();
+    assert.throws(function(){
+      nJwt.verify(header+".notavalidbody");
+    },properties.errors.PARSE_ERROR);
+  });
+
 });
 
 describe('Verifier().verify() ',function(){
